move netwotk types to helpers
