+++ src/containers/NetworkPanel/index.tsx
@@ -1,9 +1,9 @@
 import { useState, useEffect, useMemo } from "react"
 import RegexParser from "regex-parser"
 import { SplitPaneLayout } from "@/components/Layout"
-import { INetworkRequest } from "@/hooks/useNetworkMonitor"
 import { onNavigate } from "@/services/networkMonitor"
 import { IWebSocketNetworkRequest } from "@/hooks/useWebSocketNetworkMonitor"
+import { INetworkRequest } from "@/helpers/networkHelpers"
 import { NetworkTable, INetworkTableDataRow } from "./NetworkTable"
 import { NetworkDetails } from "./NetworkDetails"
 import { Toolbar } from "../Toolbar"
